ar0330_from_ar0543: change s32 to s64 for LONG_MAX
Fixes the following build warning:

        atomisp-ar0330_from_ar0543_raw/ar0543_raw.c: In function ‘nearest_resolution_index’:
        include/linux/kernel.h:24:25: warning: overflow in conversion from ‘long int’ to ‘s32’ {aka ‘int’} changes value from ‘9223372036854775807’ to ‘-1’ [-Woverflow]
           24 | #define LONG_MAX        ((long)(~0UL>>1))
              |                         ^
        atomisp-ar0330_from_ar0543_raw/ar0543_raw.c:1366:21: note: in expansion of macro ‘LONG_MAX’
         1366 |                 m = LONG_MAX;
              |                     ^~~~~~~~

+++ atomisp-ar0330_from_ar0543_raw/ar0543_raw.c
@@ -1362,7 +1362,7 @@ static int nearest_resolution_index(int w, int h)
 	int dist;
 	int min_dist = INT_MAX;
 	struct ar0543_raw_resolution *tmp_res = NULL;
-	s32 m = LARGEST_ALLOWED_RATIO_MISMATCH;
+	s64 m = LARGEST_ALLOWED_RATIO_MISMATCH;
 
 	for (j = 0; j < 2; ++j) {
 		for (i = 0; i < N_RES; i++) {
